apps/rmbg/pyproject.toml
[project]
name = "rmbg"
version = "0.1.0"
description = "CLI background removal tool"
license = { text = "Proprietary" }
requires-python = ">=3.10"
dependencies = [
    "cryptography>=45.0.0",
    "einops>=0.8.2",
    "huggingface-hub>=1.8.0",
    "kornia>=0.8.2",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=4.39.1",
]

[project.scripts]
rmbg = "rmbg_cli.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
build = [
    "pyinstaller>=6.16.0",
]
